Move debugging session into LogicScriptDebugger
