+++ LogicScript/Extensions.cs
@@ -9,8 +9,3 @@ namespace LogicScript
             => new(context);
     }
 }
-
-namespace System.Runtime.CompilerServices
-{
-    internal static class IsExternalInit { }
-}
